+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 11.20.2016 **[v0.01.02]**
+### 21.11.2016 [*v0.01.03*]
+
+ - Moved color manipulating methods into new ColorUtils class
+ - Improved appearance of widgets (added radial gradient and indicator shadow)
+
+### 20.11.2016 [*v0.01.02*]
 
  - Completed documentation/commenting
  - Modified Prefab and ActiveItem to implement new 'Configurable' interface, merging shared methods
@@ -7,6 +12,6 @@
  - Refactored ColorPickerView to ColorSwitcherView to avoid conflict with library class
  - Modified ColorSwitcherView to allow selection of a specific color
  
-## 11.19.2016 **[v0.01.01]**
+### 19.11.2016 [*v0.01.01*]
  
  - Initial release